feat: implementa ManutencaoService

+++ src/main/java/com/gestaooserp/dev/repository/ManutencaoItemRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.ManutencaoItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManutencaoItemRepository extends JpaRepository<ManutencaoItem,Integer> {
+public interface ManutencaoItemRepository extends JpaRepository<ManutencaoItem,Long> {
 }
