+++ src/main/java/com/gestaooserp/dev/repository/ManutencaoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.Manutencao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManutencaoRepository extends JpaRepository<Manutencao,Integer> {
+public interface ManutencaoRepository extends JpaRepository<Manutencao,Long> {
 }
